+++ seedfarmer/__init__.py
@@ -94,7 +94,7 @@ class Config(object):
         def configure(configuration: CodeSeederConfig) -> None:
             LOGGER.debug(f"OPS ROOT (OPS_ROOT) is {self.OPS_ROOT}")
             configuration.timeout = 120
-            configuration.codebuild_image = "public.ecr.aws/v3o4w1g6/aws-codeseeder/code-build-base:2.2.0"
+            configuration.codebuild_image = "public.ecr.aws/v3o4w1g6/aws-codeseeder/code-build-base:3.0.0"
             configuration.pre_build_commands = [
                 (
                     "nohup /usr/local/bin/dockerd --host=unix:///var/run/docker.sock"